+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.0.0 (Apr 20, 2021)
+
+- **[Feature]**: Updated to react-native 0.64.0
+- **[Fix]**: Updated npm dependencies
+- **[Fix]**: Fixed react-native 0.64.0 `nativeEvents.listeners()` crash [PR485](https://github.com/opentok/opentok-react-native/pull/485)
+
+# 0.16.0 (Apr 19, 2021)
+
+- **[Feature]**: Added support for `stereo` output on `iOS` and `Android` using `CustomAudioDevice`
+
 # 0.15.0 (Jan 26, 2021)
 
 - **[Feature]**: OTSubscriber: added [preferredResolution](https://tokbox.com/developer/sdks/js/reference/Subscriber.html#setPreferredResolution) and [preferredFrameRate](https://tokbox.com/developer/sdks/js/reference/Subscriber.html#setPreferredFrameRate) properties.